backend/requirements.txt
# Core ML
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
timm==0.9.12

# Data Processing
numpy==1.24.3
pandas==2.0.3
Pillow==10.1.0
opencv-python==4.8.1.78
albumentations==1.3.1

# NLP Metrics
nltk==3.8.1
pycocoevalcap==1.2

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Storage & Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Monitoring & Logging
loguru==0.7.2

# Security
bcrypt==4.1.1
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Deployment Scripts Dependencies
questionary==2.0.1
rich==13.7.0
tabulate==0.9.0
